compose the typography, reducing lines a lot, as well as move useClipboard over
